Backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5.0
python-dotenv>=1.1,<2.0
